[REFACTOR] Move align.aligned into align.align
